% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedStand.R
\name{mixedStand}
\alias{mixedStand}
\title{Standardize inputs for fair variable selection in sparse penalized regression.}
\usage{
mixedStand(
  x,
  y,
  standardization = "Mixed",
  continuousVars = NULL,
  binaryVars = NULL,
  categoricalVars = NULL,
  pDichoto = 0.5,
  family = "binomial"
)
}
\arguments{
\item{x}{an input data frame.}

\item{y}{outcome variable.}

\item{standardization}{a character of specified standardization methods. Users can choose from 5 methods:
`Zscore`, `Gelman`, `Minmax`, `Bring`, `Mixed`. The default is the `Mixed` method.}

\item{continuousVars}{either a character vector of continuous variable names, or a
numeric vector of continuous variable column indices.}

\item{binaryVars}{either a character vector of binary variable names, or a
numeric vector of binary variable column indices.}

\item{categoricalVars}{either a character vector of categorical variable names, or a
numeric vector of categorical variable column indices.}

\item{pDichoto}{threshold of converting continuous variable to binary variable.}

\item{family}{can be "binomial" for logistic regression and "gaussian" for linear regression.}
}
\value{
Standardized data.
\item{xStand}{standardized data}
}
\description{
`mixedStand( )` standardize variables using specified standardization method.
}
\author{
Xiang Li, Qing Pan, Yong Ma
}
